Trip transitions call Payment methods
I added some hooks on the Trip states so setting or transitionning to a state calls aPayment method :
- Created calls Payment.bill
- Started calls Payment.pay
- Cancelled calls Payment.reimburse

It is important to note that these methods are called when the state is changed through the normal flow of aasm transitions. So `Trip.find_by(...).start` will call `Payment.pay`, but `Trip.create(state: "started")` will not.
I added some tests to document this behavior.

In the future, we may want to add the `no_direct_assignment: true` aasm option to ensure there is no direct assignment of the Trips state, and that these are set only by using aas transitions flow.