Setup rabbitmq
Added a new container 'rabbitmq' in docker-compose. Rabbitmq is a message queuing server we will use to publish Trips state changes.
To the gemfile I added the 'bunny' gem which provide a rabbitmq client for Ruby.

I also added a new lib called "Publisher". It allows to easily publish a message to the rabbitmq via its ::publish method. The message here is a json encoded string.