Add publisher test
Added a test for the Publisher lib.
I had to put a `sleep(0.5)` to wait for the message to be fired to subscripbers, but I don't like this as it is not robust : if the message takes longer to be sent, the test will fail. If you have a better solution I'll take it.